Fix compilation warnings on Windows
work-items: #62916
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C63679

+++ development/libraries/DataAccess/src/FaultElementCalculator.C
@@ -245,8 +245,8 @@ void FaultElementCalculator::computeGridIntersectionPoints ( const LineSegment& 
                                                                    bool&          translatedNeeded ) const
 {
    int I;
-   int oldSize;
-   int newSize;
+   size_t oldSize;
+   size_t newSize;
 
    Point P;
    PointLessThan PLT;
